third_party/WebKit/LayoutTests/media/remoteplayback/watch-availability-works-when-backend-disabled.html

<!DOCTYPE html>
<title>Test that watchAvailability does not throw when backend is disabled.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gc.js"></script>
<script src="../media-file.js"></script>
<script src="util.js"></script>
<video></video>
<script>
  // WPT: this test can't be moved to WPT because of its usage of:
  // - internals.runtimeFlags

  async_test(function(t) {
    disableRemotePlaybackBackendForTest(t);

    var availability_callback_id = -1;
    var v = document.querySelector('video');
    v.src = findMediaFile('video', 'content/test');

    function callback(available) {
      assert_false(available);
      v.remote.cancelWatchAvailability(availability_callback_id).then(
         t.step_func_done(), t.unreached_func());
    }

    v.remote.watchAvailability(callback).then(t.step_func(
        function(id) { assert_true(id >= 0); availability_callback_id = id; }),
        t.unreached_func());
  }, 'Test that watchAvailability doesn not throw when backend is disabled.');
</script>
